requirements.txt
requests
pandas
python-dotenv
google-genai
PyPDF2
ntplib
supabase==2.3.4
rich==13.7.0
